refactor: change paths to constants
